Reworked the site routing.
