+++ src/component/home-page.js
@@ -1,4 +1,3 @@
-import { TitlePageLayout } from './title-page-layout.js';
 import { BlogList } from './blog-list.js';
 
 const template = document.createElement('template');
@@ -12,9 +11,6 @@ export class HomePage extends HTMLElement {
   }
 
   connectedCallback() {
-    if (customElements.get('cw-title-page-layout') === undefined) {
-      customElements.define('cw-title-page-layout', TitlePageLayout);
-    }
     if (customElements.get('cw-blog-list') === undefined) {
       customElements.define('cw-blog-list', BlogList);
     }
